Script updates, tasks, version, renames

+++ tools/ci/test.sh
@@ -8,18 +8,16 @@ uc_lib=script/user-conf
 . "$uc_lib"/lib.sh
 
 mkdir -vp build
-test ! -e ./build/test-results.tap ||
-  rm ./build/test-results.tap
-
-log "Hostname: $(hostname)"
+test ! -e ./build/test-results.tap || rm ./build/test-results.tap
 
 exec 3> ./build/test-results.tap
-uc__test "$@" 1>&3 || result=$?
+#uc__test "$@" 1>&3 || result=$?
+bats test/*-spec.bats 1>&3 || result=$?
 exec 3<&-
 
-test ! -s build/test-results.tap || {
+test ! -s ./build/test-results.tap || {
   log "Test results:"
-  cat build/test-results.tap
+  cat ./build/test-results.tap | script/bats-colorize.sh
 }
 
 test -n "$result" -o "$result" = "0" &&
@@ -27,4 +25,3 @@ test -n "$result" -o "$result" = "0" &&
   log "Test OK"
 
 exit $result
-
